+++ CHANGELOG.md
@@ -1,3 +1,27 @@
 # 1.0.0
 
 Initial release of cpio-downloads-server
+
+# 1.0.1
+
+Additional attribute to configure WebUI port for NZBGet
+Additional attribute to configure WebUI port for qBittorrent
+Additional attribute to configure WebUI port for Sonarr
+Additional attribute to configure WebUI port for CouchPotato
+Additional attribute to configure Incoming Torrent port for qBittorrent
+
+# 1.0.2
+
+Added Attribute for Downloads Directory
+Added Samba Share to Downloads Directory with guest access
+Configured NZBGet and qBittorrent to use Downloads Directory by default
+
+# 1.0.3
+
+Removed "My User" attributes and user, just login as root
+Updated Fish Shell replacement to from "My User" to root
+Updated Fish Attribute from ['cpio-downloads-server'] to ['fish']
+
+# 1.0.4
+Updated Readme.md file
+Updated qBittorrent Config File to support version 3.3.3.